requirements.txt
ipykernel==6.29.5
sentence-transformers==3.2.1
langchain==0.3.7
fastapi==0.115.4
uvicorn==0.32.0
pymupdf==1.24.13
python-dotenv==1.0.1
pinecone[grpc]==5.3.1
langchain-pinecone==0.2.0
langchain-community==0.3.5
langchain-openai==0.2.6
langchain-experimental==0.3.3
pypdf==5.1.0
langchain-groq==0.2.1
python-multipart==0.0.17
pytest==8.3.3
pytest-asyncio==0.24.0
websockets==13.1
psycopg2-binary==2.9.10
slowapi==0.1.9
streamlit
-e .